+++ ChangeLog.md
@@ -6,6 +6,7 @@ ChangeLog
 
 ### Changed
 
+* Library: Mailbox: pop: Make it possible to set a flexible timeout value.
 * Unit test: Use chrono_literals.
 
 [0.0.0.2024041101] - 2024-04-11